change customerId to int, fix https thing, can do getall on paymenttype
