+++ BangazonAPI/Controllers/CustomersController.cs
@@ -1,190 +1,190 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data;
-using System.Data.SqlClient;
-using System.Linq;
-using System.Threading.Tasks;
-using BangazonAPI.Models;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Configuration;
+﻿//using System;
+//using System.Collections.Generic;
+//using System.Data;
+//using System.Data.SqlClient;
+//using System.Linq;
+//using System.Threading.Tasks;
+//using BangazonAPI.Models;
+//using Microsoft.AspNetCore.Http;
+//using Microsoft.AspNetCore.Mvc;
+//using Microsoft.Extensions.Configuration;
 
-namespace BangazonAPI.Controllers
-{
-    [Route("api/[controller]")]
-    [ApiController]
-    public class CustomersController : ControllerBase
-    {
-        private readonly IConfiguration _config;
+//namespace BangazonAPI.Controllers
+//{
+//    [Route("api/[controller]")]
+//    [ApiController]
+//    public class CustomersController : ControllerBase
+//    {
+//        private readonly IConfiguration _config;
 
-        public CustomersController(IConfiguration config)
-        {
-            _config = config;
-        }
+//        public CustomersController(IConfiguration config)
+//        {
+//            _config = config;
+//        }
 
-        private SqlConnection Connection
-        {
-            get
-            {
-                return new SqlConnection(_config.GetConnectionString("DefaultConnection"));
-            }
-        }
+//        private SqlConnection Connection
+//        {
+//            get
+//            {
+//                return new SqlConnection(_config.GetConnectionString("DefaultConnection"));
+//            }
+//        }
 
-        // GET api/values
-        [HttpGet]
-        public async Task<IActionResult> Get()
-        {
-            using (SqlConnection conn = Connection)
-            {
-                conn.Open();
-                using (SqlCommand cmd = conn.CreateCommand())
-                {
-                    cmd.CommandText = "Write your SQL statement here to get all customers";
-                    SqlDataReader reader = await cmd.ExecuteReaderAsync();
+//        // GET api/values
+//        [HttpGet]
+//        public async Task<IActionResult> Get()
+//        {
+//            using (SqlConnection conn = Connection)
+//            {
+//                conn.Open();
+//                using (SqlCommand cmd = conn.CreateCommand())
+//                {
+//                    cmd.CommandText = "Write your SQL statement here to get all customers";
+//                    SqlDataReader reader = await cmd.ExecuteReaderAsync();
 
-                    List<Customer> customers = new List<Customer>();
-                    while (reader.Read())
-                    {
-                        Customer customer = new Customer
-                        {
-                            Id = reader.GetInt32(reader.GetOrdinal("Id")),
-                            FirstName = reader.GetString(reader.GetOrdinal("FirstName")),
-                            LastName = reader.GetString(reader.GetOrdinal("LastName")),
-                            // You might have more columns
-                        };
+//                    List<Customer> customers = new List<Customer>();
+//                    while (reader.Read())
+//                    {
+//                        Customer customer = new Customer
+//                        {
+//                            Id = reader.GetInt32(reader.GetOrdinal("Id")),
+//                            FirstName = reader.GetString(reader.GetOrdinal("FirstName")),
+//                            LastName = reader.GetString(reader.GetOrdinal("LastName")),
+//                            // You might have more columns
+//                        };
 
-                        customers.Add(customer);
-                    }
+//                        customers.Add(customer);
+//                    }
 
-                    reader.Close();
+//                    reader.Close();
 
-                    return Ok(customers);
-                }
-            }
-        }
+//                    return Ok(customers);
+//                }
+//            }
+//        }
 
-        // GET api/values/5
-        [HttpGet("{id}")]
-        public async Task<IActionResult> Get(int id)
-        {
-            using (SqlConnection conn = Connection)
-            {
-                conn.Open();
-                using (SqlCommand cmd = conn.CreateCommand())
-                {
-                    cmd.CommandText = "Write your SQL statement here to get a single customer";
-                    cmd.Parameters.Add(new SqlParameter("@id", id));
-                    SqlDataReader reader = await cmd.ExecuteReaderAsync();
+//        // GET api/values/5
+//        [HttpGet("{id}")]
+//        public async Task<IActionResult> Get(int id)
+//        {
+//            using (SqlConnection conn = Connection)
+//            {
+//                conn.Open();
+//                using (SqlCommand cmd = conn.CreateCommand())
+//                {
+//                    cmd.CommandText = "Write your SQL statement here to get a single customer";
+//                    cmd.Parameters.Add(new SqlParameter("@id", id));
+//                    SqlDataReader reader = await cmd.ExecuteReaderAsync();
 
-                    Customer customer = null;
-                    if (reader.Read())
-                    {
-                        customer = new Customer
-                        {
-                            Id = reader.GetInt32(reader.GetOrdinal("Id")),
-                            FirstName = reader.GetString(reader.GetOrdinal("FirstName")),
-                            LastName = reader.GetString(reader.GetOrdinal("LastName")),
-                            // You might have more columns
-                        };
-                    }
+//                    Customer customer = null;
+//                    if (reader.Read())
+//                    {
+//                        customer = new Customer
+//                        {
+//                            Id = reader.GetInt32(reader.GetOrdinal("Id")),
+//                            FirstName = reader.GetString(reader.GetOrdinal("FirstName")),
+//                            LastName = reader.GetString(reader.GetOrdinal("LastName")),
+//                            // You might have more columns
+//                        };
+//                    }
 
-                    reader.Close();
+//                    reader.Close();
 
-                    return Ok(customer);
-                }
-            }
-        }
+//                    return Ok(customer);
+//                }
+//            }
+//        }
 
-        // POST api/values
-        [HttpPost]
-        public async Task<IActionResult> Post([FromBody] Customer customer)
-        {
-            using (SqlConnection conn = Connection)
-            {
-                conn.Open();
-                using (SqlCommand cmd = conn.CreateCommand())
-                {
-                    // More string interpolation
-                    cmd.CommandText = @"
-                        INSERT INTO Customer ()
-                        OUTPUT INSERTED.Id
-                        VALUES ()
-                    ";
-                    cmd.Parameters.Add(new SqlParameter("@firstName", customer.FirstName));
+//        // POST api/values
+//        [HttpPost]
+//        public async Task<IActionResult> Post([FromBody] Customer customer)
+//        {
+//            using (SqlConnection conn = Connection)
+//            {
+//                conn.Open();
+//                using (SqlCommand cmd = conn.CreateCommand())
+//                {
+//                    // More string interpolation
+//                    cmd.CommandText = @"
+//                        INSERT INTO Customer ()
+//                        OUTPUT INSERTED.Id
+//                        VALUES ()
+//                    ";
+//                    cmd.Parameters.Add(new SqlParameter("@firstName", customer.FirstName));
 
-                    customer.Id = (int) await cmd.ExecuteScalarAsync();
+//                    customer.Id = (int) await cmd.ExecuteScalarAsync();
 
-                    return CreatedAtRoute("GetCustomer", new { id = customer.Id }, customer);
-                }
-            }
-        }
+//                    return CreatedAtRoute("GetCustomer", new { id = customer.Id }, customer);
+//                }
+//            }
+//        }
 
-        // PUT api/values/5
-        [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, [FromBody] Customer customer)
-        {
-            try
-            {
-                using (SqlConnection conn = Connection)
-                {
-                    conn.Open();
-                    using (SqlCommand cmd = conn.CreateCommand())
-                    {
-                        cmd.CommandText = @"
-                            UPDATE Customer
-                            SET FirstName = @firstName
-                            -- Set the remaining columns here
-                            WHERE Id = @id
-                        ";
-                        cmd.Parameters.Add(new SqlParameter("@id", customer.Id));
-                        cmd.Parameters.Add(new SqlParameter("@firstName", customer.FirstName));
+//        // PUT api/values/5
+//        [HttpPut("{id}")]
+//        public async Task<IActionResult> Put(int id, [FromBody] Customer customer)
+//        {
+//            try
+//            {
+//                using (SqlConnection conn = Connection)
+//                {
+//                    conn.Open();
+//                    using (SqlCommand cmd = conn.CreateCommand())
+//                    {
+//                        cmd.CommandText = @"
+//                            UPDATE Customer
+//                            SET FirstName = @firstName
+//                            -- Set the remaining columns here
+//                            WHERE Id = @id
+//                        ";
+//                        cmd.Parameters.Add(new SqlParameter("@id", customer.Id));
+//                        cmd.Parameters.Add(new SqlParameter("@firstName", customer.FirstName));
 
-                        int rowsAffected = await cmd.ExecuteNonQueryAsync();
+//                        int rowsAffected = await cmd.ExecuteNonQueryAsync();
 
-                        if (rowsAffected > 0)
-                        {
-                            return new StatusCodeResult(StatusCodes.Status204NoContent);
-                        }
+//                        if (rowsAffected > 0)
+//                        {
+//                            return new StatusCodeResult(StatusCodes.Status204NoContent);
+//                        }
 
-                        throw new Exception("No rows affected");
-                    }
-                }
-            }
-            catch (Exception)
-            {
-                if (!CustomerExists(id))
-                {
-                    return NotFound();
-                }
-                else
-                {
-                    throw;
-                }
-            }
-        }
+//                        throw new Exception("No rows affected");
+//                    }
+//                }
+//            }
+//            catch (Exception)
+//            {
+//                if (!CustomerExists(id))
+//                {
+//                    return NotFound();
+//                }
+//                else
+//                {
+//                    throw;
+//                }
+//            }
+//        }
 
-        // DELETE api/values/5
-        [HttpDelete("{id}")]
-        public async Task<IActionResult> Delete(int id)
-        {
-        }
+//        // DELETE api/values/5
+//        [HttpDelete("{id}")]
+//        public async Task<IActionResult> Delete(int id)
+//        {
+//        }
 
-        private bool CustomerExists(int id)
-        {
-            using (SqlConnection conn = Connection)
-            {
-                conn.Open();
-                using (SqlCommand cmd = conn.CreateCommand())
-                {
-                    // More string interpolation
-                    cmd.CommandText = "SELECT Id FROM Customer WHERE Id = @id";
-                    cmd.Parameters.Add(new SqlParameter("@id", id));
+//        private bool CustomerExists(int id)
+//        {
+//            using (SqlConnection conn = Connection)
+//            {
+//                conn.Open();
+//                using (SqlCommand cmd = conn.CreateCommand())
+//                {
+//                    // More string interpolation
+//                    cmd.CommandText = "SELECT Id FROM Customer WHERE Id = @id";
+//                    cmd.Parameters.Add(new SqlParameter("@id", id));
 
-                    SqlDataReader reader = cmd.ExecuteReader();
+//                    SqlDataReader reader = cmd.ExecuteReader();
 
-                    return reader.Read();
-                }
-            }
-        }
-    }
-}
+//                    return reader.Read();
+//                }
+//            }
+//        }
+//    }
+//}
